Change bitfields from enum to uint64
Implements spec change https://github.com/open-telemetry/opamp-spec/pull/125

In some languages enum fields are strongly typed and it is impossible to assigned OR-ed values to the enum field. This makes impossible to compose the bit fields properly.

This changes all bit field declarations from enum to uint64. The enum declarations stay, so that bit definitions are clear.

+++ client/clientimpl_test.go
@@ -732,7 +732,7 @@ func TestReportAgentDescription(t *testing.T) {
 			// Ask client for full AgentDescription.
 			return &protobufs.ServerToAgent{
 				InstanceUid: msg.InstanceUid,
-				Flags:       protobufs.ServerToAgentFlags_ServerToAgentFlags_ReportFullState,
+				Flags:       uint64(protobufs.ServerToAgentFlags_ServerToAgentFlags_ReportFullState),
 			}
 		})
 
@@ -805,7 +805,7 @@ func TestReportAgentHealth(t *testing.T) {
 			// Ask client for full AgentDescription.
 			return &protobufs.ServerToAgent{
 				InstanceUid: msg.InstanceUid,
-				Flags:       protobufs.ServerToAgentFlags_ServerToAgentFlags_ReportFullState,
+				Flags:       uint64(protobufs.ServerToAgentFlags_ServerToAgentFlags_ReportFullState),
 			}
 		})
 
@@ -874,7 +874,7 @@ func TestReportEffectiveConfig(t *testing.T) {
 			// Ask client for full AgentDescription.
 			return &protobufs.ServerToAgent{
 				InstanceUid: msg.InstanceUid,
-				Flags:       protobufs.ServerToAgentFlags_ServerToAgentFlags_ReportFullState,
+				Flags:       uint64(protobufs.ServerToAgentFlags_ServerToAgentFlags_ReportFullState),
 			}
 		})
 
@@ -979,7 +979,7 @@ func verifyRemoteConfigUpdate(t *testing.T, successCase bool, expectStatus *prot
 			return &protobufs.ServerToAgent{
 				InstanceUid: msg.InstanceUid,
 				// Ask client to report full status.
-				Flags: protobufs.ServerToAgentFlags_ServerToAgentFlags_ReportFullState,
+				Flags: uint64(protobufs.ServerToAgentFlags_ServerToAgentFlags_ReportFullState),
 			}
 		})
 
@@ -1169,7 +1169,7 @@ func verifyUpdatePackages(t *testing.T, testCase packageTestCase) {
 
 				if compressedReceived {
 					// Ask for full report again.
-					response.Flags = protobufs.ServerToAgentFlags_ServerToAgentFlags_ReportFullState
+					response.Flags = uint64(protobufs.ServerToAgentFlags_ServerToAgentFlags_ReportFullState)
 				} else {
 					// Keep triggering status report by setting AgentDescription
 					// until the compressed PackageStatuses arrives.
